feat: 생년월일 포맷팅 유틸함수
